+++ lib/interactive_sort.dart
@@ -1,3 +1,4 @@
 library interactive_sort;
 
 export 'src/interactive_sort_interface.dart';
+export 'src/choice_pair.dart';